Transactions - create itinerary - steps 7-10 bug fix

+++ payana/payana_service/controller/payana_bigtable_controller/payana_itinerary_objects_controller.py
@@ -368,8 +368,11 @@ class PayanaItineraryObjectDeleteTransactionEndPoint(Resource):
             pass
 
         # Delete all excursions in itinerary
-        excursion_id_list = payana_existing_itinerary_obj[
-            bigtable_constants.payana_itinerary_column_family_excursion_id_list]
+        excursion_id_list = {}
+
+        if bigtable_constants.payana_itinerary_column_family_excursion_id_list in payana_existing_itinerary_obj:
+            excursion_id_list = payana_existing_itinerary_obj[
+                bigtable_constants.payana_itinerary_column_family_excursion_id_list]
 
         for _, excursion_id in excursion_id_list.items():
 
@@ -391,16 +394,17 @@ class PayanaItineraryObjectDeleteTransactionEndPoint(Resource):
                     bigtable_constants.payana_excursion_objects_delete_failure_message, payana_itinerary_objects_name_space)
 
             # Step 3 - Delete all checkin objects
-            checkin_id_list = payana_excursion_existing_obj[
-                bigtable_constants.payana_excursion_column_family_checkin_id_list]
+            if bigtable_constants.payana_excursion_column_family_checkin_id_list in payana_excursion_existing_obj:
+                checkin_id_list = payana_excursion_existing_obj[
+                    bigtable_constants.payana_excursion_column_family_checkin_id_list]
 
-            for _, checkin_id in checkin_id_list.items():
-                delete_checkin_object_status = delete_checkin_object(
-                    checkin_id)
+                for _, checkin_id in checkin_id_list.items():
+                    delete_checkin_object_status = delete_checkin_object(
+                        checkin_id)
 
-                if not delete_checkin_object_status:
-                    # Add logging to auto-handle later
-                    pass
+                    if not delete_checkin_object_status:
+                        # Add logging to auto-handle later
+                        pass
 
             # Step 4 - Update Itinerary object by removing the excursion ID
             payana_itinerary_delete_obj = {bigtable_constants.payana_itinerary_column_family_excursion_id_list: {payana_excursion_existing_obj[bigtable_constants.payana_excursion_metadata][
